+++ CHANGELOG.md
@@ -6,6 +6,17 @@ Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how 
 
 ## [Unreleased]
 
+## [0.0.2] - 2023-11-16
+
+### Added
+
+- More customization settings:
+
+  - Search by Github type
+  - Wrap search text in quotes
+  - Limit search to one or more repositories
+  - Limit search to one or more languages
+
 ## [0.0.1] - 2023-11-16
 
 ### Added